Rename env vars and fix scripts
