+++ packages/admin-panel-server/src/index.ts
@@ -14,8 +14,8 @@ import { createApp } from './app';
 configureWinston();
 
 configureDotEnv([
-  path.resolve(__dirname, '../../../env/.env.servers'),
-  path.resolve(__dirname, '../../../env/.env.db'),
+  path.resolve(__dirname, '../../../env/servers.db'),
+  path.resolve(__dirname, '../../../env/db.env'),
   path.resolve(__dirname, '.env'),
 ]);
