+++ packages/central-server/src/configureEnv.js
@@ -7,14 +7,14 @@ import path from 'path';
 import { configureDotEnv } from '@tupaia/server-utils';
 
 const envFilePaths = [
-  path.resolve(__dirname, '../../../env/.env.db'),
-  path.resolve(__dirname, '../../../env/.env.servers'),
-  path.resolve(__dirname, '../../../env/.env.pg'),
-  path.resolve(__dirname, '../../../env/.env.dhis'),
-  path.resolve(__dirname, '../../../env/.envexternalDBConnections'),
-  path.resolve(__dirname, '../../../env/.env.mail'),
-  path.resolve(__dirname, '../../../env/.env.aws'),
-  path.resolve(__dirname, '../../../env/.env.aggregation'),
+  path.resolve(__dirname, '../../../env/db.env'),
+  path.resolve(__dirname, '../../../env/servers.db'),
+  path.resolve(__dirname, '../../../env/pg.env'),
+  path.resolve(__dirname, '../../../env/dhis.env'),
+  path.resolve(__dirname, '../../../env/external-db-connections.env'),
+  path.resolve(__dirname, '../../../env/mail.env'),
+  path.resolve(__dirname, '../../../env/aws.env'),
+  path.resolve(__dirname, '../../../env/aggregation.env'),
   path.resolve(__dirname, '.env'),
 ];
