+++ packages/database/scripts/dumpDatabase.sh
@@ -29,7 +29,8 @@ function show_loading_spinner() {
     echo "" # reset prompt
 }
 
-source ".env"
+DIR=$(pwd "$0")
+source $DIR/../../scripts/bash/mergeEnvForDB.sh 
 
 DUMP_FILE_NAME="dump.sql"
