+++ packages/database/src/configureEnv.js
@@ -9,8 +9,8 @@ const dotenv = require('dotenv');
 export const configureEnv = () => {
   dotenv.config({
     path: [
-      path.resolve(__dirname, '../../env/.env.db'),
-      path.resolve(__dirname, '../../env/.env.pg'),
+      path.resolve(__dirname, '../../../env/db.env'),
+      path.resolve(__dirname, '../../../env/pg.env'),
       path.resolve(__dirname, '.env'),
     ],
   });
