+++ packages/database/tsconfig.json
@@ -5,7 +5,6 @@
   },
   "include": [
     "src/**/*.js",
-    "configureEnv.js"
   ],
   "exclude": [
     "src/__tests__",
