+++ packages/entity-server/src/configureEnv.ts
@@ -8,7 +8,7 @@ import { configureDotEnv } from '@tupaia/server-utils';
 
 export const configureEnv = () =>
   configureDotEnv([
-    path.resolve(__dirname, '../../../env/.env.servers'),
-    path.resolve(__dirname, '../../../env/.env.db'),
+    path.resolve(__dirname, '../../../env/servers.db'),
+    path.resolve(__dirname, '../../../env/db.env'),
     path.resolve(__dirname, '.env'),
   ]); // Load the environment variables into process.env
