+++ packages/meditrak-app-server/.env.example
@@ -1,5 +1,7 @@
 PORT=
 
+API_CLIENT_SALT=
 API_CLIENT_NAME=
-API_CLIENT_PASSWORD= 
- 
+API_CLIENT_PASSWORD=
+
+AUTH_API_URL=
