+++ packages/psss-server/src/index.ts
@@ -12,8 +12,8 @@ import { createApp } from './app';
 
 configureWinston();
 configureDotEnv([
-  path.resolve(__dirname, '../../../env/.env.servers'),
-  path.resolve(__dirname, '../../../env/.env.db'),
+  path.resolve(__dirname, '../../../env/servers.db'),
+  path.resolve(__dirname, '../../../env/db.env'),
   '.env',
 ]); // Load the environment variables into process.env
