+++ packages/report-server/configureEnv.js
@@ -9,12 +9,12 @@ const dotenv = require('dotenv');
 () => {
   dotenv.config({
     path: [
-      path.resolve(__dirname, '../../../env/.env.aggregation'),
-      path.resolve(__dirname, '../../../env/.env.dhis'),
-      path.resolve(__dirname, '../../../env/.env.dataLake'),
-      path.resolve(__dirname, '../../../env/.env.superset'),
-      path.resolve(__dirname, '../../../env/.env.servers'),
-      path.resolve(__dirname, '../../../env/.env.db'),
+      path.resolve(__dirname, '../../../env/aggregation.env'),
+      path.resolve(__dirname, '../../../env/dhis.env'),
+      path.resolve(__dirname, '../../../env/data-lake.env'),
+      path.resolve(__dirname, '../../../env/superset.env'),
+      path.resolve(__dirname, '../../../env/servers.db'),
+      path.resolve(__dirname, '../../../env/db.env'),
       '.env',
     ],
   });
