+++ packages/report-server/src/index.ts
@@ -14,12 +14,12 @@ import { createApp } from './app';
 
 configureWinston();
 configureDotEnv([
-  path.resolve(__dirname, '../../../env/.env.aggregation'),
-  path.resolve(__dirname, '../../../env/.env.dhis'),
-  path.resolve(__dirname, '../../../env/.env.dataLake'),
-  path.resolve(__dirname, '../../../env/.env.superset'),
-  path.resolve(__dirname, '../../../env/.env.servers'),
-  path.resolve(__dirname, '../../../env/.env.db'),
+  path.resolve(__dirname, '../../../env/aggregation.env'),
+  path.resolve(__dirname, '../../../env/dhis.env'),
+  path.resolve(__dirname, '../../../env/data-lake.env'),
+  path.resolve(__dirname, '../../../env/superset.env'),
+  path.resolve(__dirname, '../../../env/servers.db'),
+  path.resolve(__dirname, '../../../env/db.env'),
   '.env',
 ]); // Load the environment variables into process.env
