+++ packages/web-config-server/.env.example
@@ -28,7 +28,6 @@ MICROSERVICE_CLIENT_USERNAME=
 PALAU_DHIS_API_URL=
 PALAU_DHIS_CLIENT_SECRET=
 PALAU_DHIS_PASSWORD=
-REACT_APP_CONFIG_SERVER_BASE_URL=
 SITE_EMAIL_ADDRESS=
 SMTP_HOST=
 SMTP_PASSWORD=
